Socrata grabber

+++ models.py
@@ -1,10 +1,11 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
 class RealEstateTransaction:
     town_name: str = None
-    sales_ratio: float = None
-    sale_amount: float = None
+    sales_ratio: Decimal = None
+    sale_amount: Decimal = None
     year: int = None
